+++ src/Tierlist/Tierlist.tsx
@@ -5,13 +5,25 @@ import {
   TierListItemsProps,
   TierListLabelProps,
   TierlistProps,
-  TierListRankProps,
+  TierListTierProps,
   RankedItems,
 } from "./Tierlist.types";
+import {
+  handleTouchEnd,
+  handleTouchMove,
+  handleTouchStart,
+} from "./Tierlist.utils";
 import { cn } from "../utils/cn";
 
 const TierListItems = (props: TierListItemsProps) => {
-  const { children, className, onUnrankDrop, rankedItems, ...rest } = props;
+  const {
+    children,
+    onTouchDrop,
+    onUnrankDrop,
+    rankedItems,
+    itemsClasses,
+    ...rest
+  } = props;
 
   function handleDragStart(e: React.DragEvent<HTMLDivElement>) {
     e.dataTransfer.setData("text/plain", e.currentTarget.dataset.id!);
@@ -19,8 +31,9 @@ const TierListItems = (props: TierListItemsProps) => {
 
   //handle body scrolling when dragging an item
   React.useEffect(() => {
-    function handleDragOver(e: DragEvent) {
-      const { clientY } = e;
+    function handleDragOver(e: DragEvent | TouchEvent) {
+      //get the clientY position of the cursor or touch
+      const clientY = "clientY" in e ? e.clientY : e.touches[0].clientY;
       const scrollSpeed = 5;
       const threshold = 100; // Distance from the edge to start scrolling
 
@@ -32,11 +45,27 @@ const TierListItems = (props: TierListItemsProps) => {
     }
 
     window.addEventListener("dragover", handleDragOver);
+    window.addEventListener("touchmove", handleDragOver);
 
     return () => {
       window.removeEventListener("dragover", handleDragOver);
+      window.removeEventListener("touchmove", handleDragOver);
     };
   }, []);
+
+  function handleTouchDrop(e: React.TouchEvent<HTMLDivElement>) {
+    const { targetTierIdx } = handleTouchEnd(e);
+
+    if (targetTierIdx === -1) return;
+
+    //provide the rank index and the item id to the parent component
+    onTouchDrop(targetTierIdx, parseInt((e.target as HTMLElement).dataset.id!));
+  }
+
+  function handleOnUnrankDrop(e: React.DragEvent<HTMLDivElement>) {
+    const draggedItemId = e.dataTransfer.getData("text/plain");
+    onUnrankDrop(parseInt(draggedItemId));
+  }
 
   if (!Array.isArray(children)) {
     return null;
@@ -46,14 +75,15 @@ const TierListItems = (props: TierListItemsProps) => {
     <div
       className={cn(
         "flex flex-wrap gap-2 select-none items-center rounded-md p-8 my-4 border border-clr-border border-dashed",
-        className
+        itemsClasses
       )}
-      onDrop={onUnrankDrop}
+      data-id="tierlist-items"
+      onDrop={handleOnUnrankDrop}
       onDragOver={(e) => e.preventDefault()}
       {...rest}
     >
       {React.Children.map(children, (child, idx) => {
-        //check if the item is already in a rank, so we can hide it from the list
+        //check if the item is already in a tier, so we can hide it from the list
         const isRanked =
           rankedItems
             .flat()
@@ -70,6 +100,9 @@ const TierListItems = (props: TierListItemsProps) => {
               }),
               draggable: true,
               onDragStart: handleDragStart,
+              onTouchStart: handleTouchStart,
+              onTouchMove: handleTouchMove,
+              onTouchEnd: handleTouchDrop,
               "data-id": idx,
             })}
           </React.Fragment>
@@ -80,11 +113,14 @@ const TierListItems = (props: TierListItemsProps) => {
 };
 
 const TierListLabel = (props: TierListLabelProps) => {
-  const { children: label, color } = props;
+  const { children: label, color, labelClasses } = props;
 
   return (
     <div
-      className="flex items-center px-2 justify-center self-stretch min-w-24 max-w-32 w-fit font-medium text-slate-700 select-none flex-1"
+      className={cn(
+        "flex items-center px-2 justify-center self-stretch min-w-24 max-w-32 w-fit font-medium text-slate-700 select-none flex-1",
+        labelClasses
+      )}
       style={{ backgroundColor: color }}
     >
       {label}
@@ -92,8 +128,8 @@ const TierListLabel = (props: TierListLabelProps) => {
   );
 };
 
-const TierListRank = (props: TierListRankProps) => {
-  const { items, onDrop } = props;
+const TierListTier = (props: TierListTierProps) => {
+  const { items, idx, onDrop, onUnrankDrop, onTouchDrop, tierClasses } = props;
 
   function handleDragOver(e: React.DragEvent<HTMLDivElement>) {
     //allow drop on this element
@@ -104,11 +140,32 @@ const TierListRank = (props: TierListRankProps) => {
     e.dataTransfer.setData("text/plain", e.currentTarget.dataset.id!);
   }
 
+  function handleTouchDrop(e: React.TouchEvent<HTMLDivElement>) {
+    const { isUnranking, targetTierIdx } = handleTouchEnd(e);
+
+    const itemId = parseInt((e.target as HTMLElement).dataset.id!);
+
+    //fire unranking callback if the item was dropped back to the list
+    if (isUnranking) {
+      onUnrankDrop(itemId);
+      return;
+    }
+
+    if (targetTierIdx === -1) return;
+
+    //provide the rank index and the item id to the parent component
+    onTouchDrop(targetTierIdx, itemId);
+  }
+
   return (
     <div
-      className="flex gap-4 p-4 min-h-16 flex-wrap flex-1 items-center"
+      className={cn(
+        "flex gap-4 p-4 min-h-24 flex-wrap flex-1 items-center",
+        tierClasses
+      )}
       onDragOver={handleDragOver}
       onDrop={onDrop}
+      data-tierlist-tier-idx={idx}
     >
       {React.Children.map(items, (child, idx) =>
         React.cloneElement(child, {
@@ -116,6 +173,9 @@ const TierListRank = (props: TierListRankProps) => {
           className: cn(child.props.className, "cursor-pointer"),
           draggable: true,
           onDragStart: handleRankedDragStart,
+          onTouchStart: handleTouchStart,
+          onTouchMove: handleTouchMove,
+          onTouchEnd: handleTouchDrop,
         })
       )}
     </div>
@@ -128,7 +188,10 @@ export const Tierlist = (props: TierlistProps) => {
     className,
     labels = defaultLabels,
     labelColors = defaultLabelColors,
-    onDrop,
+    onTierDrop,
+    itemsClasses,
+    labelClasses,
+    tierClasses,
     ...rest
   } = props;
 
@@ -150,11 +213,8 @@ export const Tierlist = (props: TierlistProps) => {
     });
   }) as React.ReactElement[];
 
-  function handleDrop(rankIdx: number, e: React.DragEvent<HTMLDivElement>) {
-    //get the data-id attribute of the dragged element
-    const draggedItemId = e.dataTransfer.getData("text/plain");
-
-    const draggedItem = childrenWithProps[parseInt(draggedItemId)];
+  function rankItems(rankIdx: number, itemId: number) {
+    const draggedItem = childrenWithProps[itemId];
 
     //if we dragged an item that is not in the list, return
     if (!draggedItem) return;
@@ -168,41 +228,50 @@ export const Tierlist = (props: TierlistProps) => {
       if (idx === rankIdx) {
         // Check if the item already exists in the target rank
         const itemExists = items.some(
-          (item) => item.props["data-id"] === parseInt(draggedItemId)
+          (item) => item.props["data-id"] === itemId
         );
         if (!itemExists) {
           return [...items, draggedItem];
         }
         return items;
       } else {
-        return items.filter(
-          (item) => item.props["data-id"] !== parseInt(draggedItemId)
-        );
+        return items.filter((item) => item.props["data-id"] !== itemId);
       }
     });
 
     //call the onDrop callback if the item was dropped into a different rank
     const currentRankIdx = rankedItems.findIndex((items) =>
-      items.some((item) => item.props["data-id"] === parseInt(draggedItemId))
+      items.some((item) => item.props["data-id"] === itemId)
     );
 
-    if (currentRankIdx !== rankIdx && onDrop) {
-      onDrop(labels[rankIdx]);
+    if (currentRankIdx !== rankIdx && onTierDrop) {
+      onTierDrop(labels[rankIdx]);
     }
 
     setRankedItems(newItems);
   }
 
-  const handleUnrankDrop = (e: React.DragEvent<HTMLDivElement>) => {
+  function handleDropInTier(
+    rankIdx: number,
+    e: React.DragEvent<HTMLDivElement>
+  ) {
+    //get the data-id attribute of the dragged element
     const draggedItemId = e.dataTransfer.getData("text/plain");
+    rankItems(rankIdx, parseInt(draggedItemId));
+  }
 
+  function handleTouchDrop(rankIdx: number, itemId: number) {
+    rankItems(rankIdx, itemId);
+  }
+
+  function handleUnrankDrop(itemId: number) {
     //remove the item from the rank
     const newItems = rankedItems.map((items) =>
-      items.filter((item) => item.props["data-id"] !== parseInt(draggedItemId))
+      items.filter((item) => item.props["data-id"] !== itemId)
     );
 
     setRankedItems(newItems);
-  };
+  }
 
   return (
     <div className={cn("text-clr-text select-none", className)} {...rest}>
@@ -213,17 +282,31 @@ export const Tierlist = (props: TierlistProps) => {
               className="flex justify-between items-center border-b border-b-clr-border last-of-type:border-b-0"
               key={label}
             >
-              <TierListLabel color={labelColors[idx]}>{label}</TierListLabel>
-              <TierListRank
-                key={`${label}-rank`}
+              <TierListLabel
+                color={labelColors[idx]}
+                labelClasses={labelClasses}
+              >
+                {label}
+              </TierListLabel>
+              <TierListTier
+                key={`${label}-tier`}
+                idx={idx}
                 items={rankedItems[idx]}
-                onDrop={handleDrop.bind(null, idx)}
+                onDrop={handleDropInTier.bind(null, idx)}
+                onTouchDrop={handleTouchDrop}
+                onUnrankDrop={handleUnrankDrop}
+                tierClasses={tierClasses}
               />
             </div>
           );
         })}
       </div>
-      <TierListItems rankedItems={rankedItems} onUnrankDrop={handleUnrankDrop}>
+      <TierListItems
+        rankedItems={rankedItems}
+        onUnrankDrop={handleUnrankDrop}
+        onTouchDrop={handleTouchDrop}
+        itemsClasses={itemsClasses}
+      >
         {childrenWithProps}
       </TierListItems>
     </div>